+++ Source/Parsek/ParsekFlight.cs
@@ -1987,6 +1987,7 @@ namespace Parsek
                 ui.DrawTimelineWindowIfOpen(windowRect);
                 ui.DrawKerbalsWindowIfOpen(windowRect);
                 ui.DrawCareerStateWindowIfOpen(windowRect);
+                ui.DrawLogisticsWindowIfOpen(windowRect);
                 ui.DrawSettingsWindowIfOpen(windowRect);
                 ui.DrawSpawnControlWindowIfOpen(windowRect);
                 ui.DrawGloopsRecorderWindowIfOpen(windowRect);